Remove unused functions and consolidate others
