+++ packages/pyright-internal/package.json
@@ -13,7 +13,8 @@
         "clean": "shx rm -rf ./dist ./out",
         "webpack:testserver": "webpack --config ./src/tests/lsp/webpack.testserver.config.js --mode=development",
         "webpack:testserver:watch": "npm run clean && webpack --config ./src/tests/lsp/webpack.testserver.config.js --mode development --watch --progress",
-        "test": "jest --forceExit",
+        "test": "npm run webpack:testserver && jest --forceExit",
+        "test:norebuild": "jest --forceExit",
         "test:coverage": "jest --forceExit --reporters=jest-junit --reporters=default --coverage --coverageReporters=cobertura --coverageReporters=html --coverageReporters=json",
         "test:imports": "jest importResolver.test --forceExit --runInBand"
     },
